#10170 move filter actions
